Improve checklist_print_overview()
Currently, teachers will always see a "You have a Checklist that needs
attention" notice in the Course Overview block. Rarely do teachers care
about their own checklist, and there I didn't take the time to handle
the logic for the check for CHECKLIST_MARKING_TEACHER and
CHECKLIST_MARKING_BOTH to see what a teacher actually needs to do with
each of their student's checklist. So this solution just hides the
"overview" alert for teachers.

The other addition was to check to see if the checklist was 100%
complete. If it is, the student shouldn't see the "Checklist needs
attention" alert anymore.